pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SICdb-MEDS"
dynamic = ["version"]
authors = [
  { name="Robin van de Water", email="robin.vandewater@hpi.de" },
]
description = "A template ETL pipeline to extract arbitrary data into the MEDS format."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "meds-transforms~=0.6.0", "requests", "beautifulsoup4", "hydra-core", "loguru", "MEDS-extract~=0.4.0"
]

[tool.setuptools_scm]

[project.optional-dependencies]
dev = ["pre-commit<4"]
tests = ["pytest", "pytest-cov"]
local_parallelism = ["hydra-joblib-launcher"]
slurm_parallelism = ["hydra-submitit-launcher"]

[project.scripts]
MEDS_extract-SICdb= "SICdb_MEDS.__main__:main"

[project.urls]
Homepage = "https://github.com/rvandewater/SICdb_MEDS"
Issues = "https://github.com/rvandewater/SICdb_MEDS/issues"
